+++ apps/prairielearn/src/pages/administratorCourses/administratorCourses.js
@@ -82,10 +82,7 @@ router.post(
     } else if (req.body.__action === 'create_course_from_request') {
       await createCourseFromRequest(req, res);
     } else {
-      throw error.make(400, 'unknown __action', {
-        locals: res.locals,
-        body: req.body,
-      });
+      throw error.make(400, `unknown __action: ${req.body.__action}`);
     }
   }),
 );