+++ apps/prairielearn/src/pages/instructorAssessmentRegrading/instructorAssessmentRegrading.js
@@ -43,12 +43,7 @@ router.post('/', function (req, res, next) {
       },
     );
   } else {
-    return next(
-      error.make(400, 'unknown __action', {
-        locals: res.locals,
-        body: req.body,
-      }),
-    );
+    return next(error.make(400, `unknown __action: ${req.body.__action}`));
   }
 });