+++ apps/prairielearn/src/pages/studentAssessmentInstance/studentAssessmentInstance.js
@@ -8,7 +8,7 @@ import { loadSqlEquiv, queryRow, queryRows } from '@prairielearn/postgres';
 
 import * as assessment from '../../lib/assessment';
 import {
-  getAssessmentPermissions,
+  canUserAssignGroupRoles,
   getGroupConfig,
   getGroupInfo,
   getQuestionGroupPermissions,
@@ -69,15 +69,16 @@ async function processFileUpload(req, res) {
   if (!res.locals.authz_result.active) {
     throw new Error(`This assessment is not accepting submissions at this time.`);
   }
-  await uploadFile(
-    req.file.originalname,
-    req.file.buffer,
-    'student_upload',
-    res.locals.assessment_instance.id,
-    null,
-    res.locals.user.user_id,
-    res.locals.authn_user.user_id,
-  );
+  await uploadFile({
+    display_filename: req.file.originalname,
+    contents: req.file.buffer,
+    type: 'student_upload',
+    assessment_id: res.locals.assessment.id,
+    assessment_instance_id: res.locals.assessment_instance.id,
+    instance_question_id: null,
+    user_id: res.locals.user.user_id,
+    authn_user_id: res.locals.authn_user.user_id,
+  });
 }
 
 async function processTextUpload(req, res) {
@@ -85,15 +86,16 @@ async function processTextUpload(req, res) {
   if (!res.locals.authz_result.active) {
     throw new Error(`This assessment is not accepting submissions at this time.`);
   }
-  await uploadFile(
-    req.body.filename,
-    Buffer.from(req.body.contents),
-    'student_upload',
-    res.locals.assessment_instance.id,
-    null,
-    res.locals.user.user_id,
-    res.locals.authn_user.user_id,
-  );
+  await uploadFile({
+    display_filename: req.body.filename,
+    contents: Buffer.from(req.body.contents),
+    type: 'student_upload',
+    assessment_id: res.locals.assessment.id,
+    assessment_instance_id: res.locals.assessment_instance.id,
+    instance_question_id: null,
+    user_id: res.locals.user.user_id,
+    authn_user_id: res.locals.authn_user.user_id,
+  });
 }
 
 async function processDeleteFile(req, res) {
@@ -123,7 +125,7 @@ router.post(
       !res.locals.authz_result.authorized_edit &&
       !res.locals.authz_data.has_course_instance_permission_edit
     ) {
-      throw error.make(403, 'Not authorized', res.locals);
+      throw error.make(403, 'Not authorized');
     }
     if (
       !res.locals.authz_result.authorized_edit &&
@@ -131,7 +133,7 @@ router.post(
         req.body.__action,
       )
     ) {
-      throw error.make(403, 'Action is only permitted to students, not staff', res.locals);
+      throw error.make(403, 'Action is only permitted to students, not staff');
     }
 
     if (req.body.__action === 'attach_file') {
@@ -160,10 +162,7 @@ router.post(
         }
         closeExam = true;
       } else {
-        throw error.make(400, 'unknown __action', {
-          locals: res.locals,
-          body: req.body,
-        });
+        throw error.make(400, `unknown __action: ${req.body.__action}`);
       }
       const requireOpen = true;
       await assessment.gradeAssessmentInstanceAsync(
@@ -200,12 +199,7 @@ router.post(
       );
       res.redirect(req.originalUrl);
     } else {
-      next(
-        error.make(400, 'unknown __action', {
-          locals: res.locals,
-          body: req.body,
-        }),
-      );
+      next(error.make(400, `unknown __action: ${req.body.__action}`));
     }
   }),
 );
@@ -265,11 +259,7 @@ router.get(
       res.locals.used_join_code = req.body.used_join_code;
 
       if (groupConfig.has_roles) {
-        const result = await getAssessmentPermissions(
-          res.locals.assessment.id,
-          res.locals.user.user_id,
-        );
-        res.locals.userCanAssignRoles = result.can_assign_roles_at_start;
+        res.locals.userCanAssignRoles = canUserAssignGroupRoles(groupInfo, res.locals.user.user_id);
 
         res.locals.user_group_roles =
           groupInfo.rolesInfo?.roleAssignments?.[res.locals.authz_data.user.uid]