+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/postgres-tools
 
+## 1.1.15
+
+### Patch Changes
+
+- 3395c25: Upgrade all JavaScript dependencies
+- Updated dependencies [3395c25]
+  - @prairielearn/postgres@1.9.2
+
+## 1.1.14
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+- Updated dependencies [dce0fa3]
+  - @prairielearn/postgres@1.9.1
+
 ## 1.1.13
 
 ### Patch Changes