+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/postgres
 
+## 1.9.2
+
+### Patch Changes
+
+- 3395c25: Upgrade all JavaScript dependencies
+
+## 1.9.1
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
 ## 1.9.0
 
 ### Minor Changes